+++ packages/cli/src/fs/index.ts
@@ -1,6 +1,4 @@
-import type WebManglerCliFile from "./file.class";
-
-export { readFilesInAll } from "./read";
+export { readFilesFiltered } from "./read";
 export { writeFiles } from "./write";
 
-export type { WebManglerCliFile };
+export type { WebManglerCliFile } from "./types";