+++ packages/cli/src/fs/write.ts
@@ -1,14 +1,18 @@
 import type { WebManglerCliFile } from "./types";
 
-import * as fs from "fs";
+import { promises as fs } from "fs";
 
 /**
  * Write a list of files to disk.
  *
  * @param files The files to write.
  */
-export function writeFiles(files: WebManglerCliFile[]): void {
-  files.forEach((file) => {
-    fs.writeFileSync(file.path, file.content);
-  });
+export async function writeFiles(files: WebManglerCliFile[]): Promise<void> {
+  const promises: Promise<void>[] = [];
+  for (const file of files) {
+    const writePromise = fs.writeFile(file.path, file.content);
+    promises.push(writePromise);
+  }
+
+  await Promise.all(promises);
 }